Moving the changes from logic base locations to side-locations, part buildMemory
