+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/gerritnotifier/LocalNotificationClaimStrategy.java
@@ -1,0 +1,47 @@
+/*
+ *  The MIT License
+ *
+ *  Copyright (c) 2010, 2014 Sony Mobile Communications Inc. All rights reserved.
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+package com.sonyericsson.hudson.plugins.gerrit.trigger.gerritnotifier;
+
+import com.sonymobile.tools.gerrit.gerritevents.dto.events.GerritTriggeredEvent;
+import edu.umd.cs.findbugs.annotations.NonNull;
+
+/**
+ * Local (non-cluster) implementation of notification claiming.
+ * Always returns true since there's no need for coordination in standalone mode.
+ *
+ * @author Robert Sandell &lt;robert.sandell@sonyericsson.com&gt;
+ */
+public class LocalNotificationClaimStrategy implements NotificationClaimStrategy {
+
+    @Override
+    public boolean tryClaimNotificationRight(@NonNull GerritTriggeredEvent event) {
+        // In local mode, always send notifications - no coordination needed
+        return true;
+    }
+
+    @Override
+    public void releaseNotificationRight(@NonNull GerritTriggeredEvent event) {
+        // No-op in local mode - nothing to release
+    }
+}
