+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/gerritnotifier/NotificationClaimStrategy.java
@@ -1,0 +1,55 @@
+/*
+ *  The MIT License
+ *
+ *  Copyright (c) 2010, 2014 Sony Mobile Communications Inc. All rights reserved.
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+package com.sonyericsson.hudson.plugins.gerrit.trigger.gerritnotifier;
+
+import com.sonymobile.tools.gerrit.gerritevents.dto.events.GerritTriggeredEvent;
+import edu.umd.cs.findbugs.annotations.NonNull;
+
+/**
+ * Strategy interface for notification claiming in different deployment modes.
+ * Implementations handle the coordination of which Jenkins instance should send
+ * notifications to Gerrit in HA/HS deployments.
+ *
+ * @author Robert Sandell &lt;robert.sandell@sonyericsson.com&gt;
+ */
+public interface NotificationClaimStrategy {
+
+    /**
+     * Attempts to claim the right to send notification for an event.
+     * In local mode, always returns true. In cluster mode, uses distributed
+     * coordination to ensure only one replica sends the notification.
+     *
+     * @param event the Gerrit event
+     * @return true if this instance should send the notification, false otherwise
+     */
+    boolean tryClaimNotificationRight(@NonNull GerritTriggeredEvent event);
+
+    /**
+     * Releases the notification claim for an event.
+     * Called after notification is sent or on error to clean up resources.
+     *
+     * @param event the Gerrit event
+     */
+    void releaseNotificationRight(@NonNull GerritTriggeredEvent event);
+}
