+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/gerritnotifier/model/BuildMemory.java
@@ -24,14 +24,7 @@
  */
 package com.sonyericsson.hudson.plugins.gerrit.trigger.gerritnotifier.model;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.map.IMap;
-import com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.BuildMemoryKey;
-import com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.EntryData;
-import com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.HazelcastInstanceProvider;
-import com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.MemoryImprintData;
+import com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.ClusterModeProvider;
 import com.sonyericsson.hudson.plugins.gerrit.trigger.diagnostics.BuildMemoryReport;
 import com.sonyericsson.hudson.plugins.gerrit.trigger.gerritnotifier.model.BuildMemory.MemoryImprint.Entry;
 import com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.GerritCause;
@@ -60,15 +53,25 @@ import static com.sonyericsson.hudson.plugins.gerrit.trigger.utils.Logic.shouldS
 /**
  * Keeps track of what builds have been triggered and if all builds are done for specific events.
  * <p>
- * Supports dual-mode operation:
- * <ul>
- *   <li><strong>Local mode</strong> (standalone): Uses in-memory TreeMap</li>
- *   <li><strong>Distributed mode</strong> (cluster): Uses Hazelcast IMap</li>
- * </ul>
+ * This is the local-mode implementation using in-memory TreeMap.
+ * For cluster mode, see {@link ClusterBuildMemory}.
  *
  * @author Robert Sandell &lt;robert.sandell@sonyericsson.com&gt;
  */
 public class BuildMemory {
+
+    /**
+     * Factory method to create the appropriate BuildMemory implementation.
+     * Returns ClusterBuildMemory if cluster mode is enabled, otherwise standard BuildMemory.
+     *
+     * @return BuildMemory instance (either local or cluster)
+     */
+    public static BuildMemory create() {
+        if (ClusterModeProvider.isClusterModeEnabled()) {
+            return new ClusterBuildMemory();
+        }
+        return new BuildMemory();
+    }
 
     /**
      * Compares GerritTriggeredEvents using the Object.hashCode() method. This ensures that every event received from
@@ -95,193 +98,12 @@ public class BuildMemory {
     private static final Logger logger = LoggerFactory.getLogger(BuildMemory.class);
 
     /**
-     * Hazelcast map name for distributed build memory.
+     * Local mode storage.
      */
-    private static final String MAP_NAME = "gerrit-trigger-build-memory";
-
-    /**
-     * Gson instance for JSON serialization of events.
-     */
-    private static final Gson GSON = new GsonBuilder().create();
-
-    /**
-     * Local mode storage (standalone Jenkins).
-     */
-    private TreeMap<GerritTriggeredEvent, MemoryImprint> memory =
+    protected TreeMap<GerritTriggeredEvent, MemoryImprint> memory =
             new TreeMap<GerritTriggeredEvent, MemoryImprint>(
                     new GerritTriggeredEventComparator());
 
-    /**
-     * Distributed mode storage (cluster mode).
-     * Lazy-initialized when first accessed in cluster mode.
-     */
-    private transient IMap<BuildMemoryKey, MemoryImprintData> distributedMemory = null;
-
-    /**
-     * Checks if cluster mode is enabled via system property.
-     *
-     * @return true if cluster mode is enabled
-     */
-    private boolean isClusterModeEnabled() {
-        return com.sonyericsson.hudson.plugins.gerrit.trigger.cluster.ClusterModeProvider.isClusterModeEnabled();
-    }
-
-    /**
-     * Gets or initializes the distributed memory map.
-     * <p>
-     * Returns null if cluster mode is disabled or Hazelcast is unavailable.
-     *
-     * @return distributed memory map, or null
-     */
-    private IMap<BuildMemoryKey, MemoryImprintData> getDistributedMemory() {
-        if (distributedMemory == null && isClusterModeEnabled()) {
-            try {
-                HazelcastInstance hz = HazelcastInstanceProvider.getInstance();
-                if (hz != null) {
-                    distributedMemory = hz.getMap(MAP_NAME);
-                    logger.debug("Initialized distributed BuildMemory map: {}", MAP_NAME);
-                }
-            } catch (Exception e) {
-                logger.warn("Failed to initialize distributed BuildMemory, falling back to local", e);
-            }
-        }
-        return distributedMemory;
-    }
-
-    /**
-     * Serializes a GerritTriggeredEvent to JSON.
-     *
-     * @param event the event to serialize
-     * @return JSON string, or null if serialization fails
-     */
-    private String serializeEvent(GerritTriggeredEvent event) {
-        try {
-            return GSON.toJson(event);
-        } catch (Exception e) {
-            logger.error("Failed to serialize event to JSON: " + event, e);
-            return null;
-        }
-    }
-
-    /**
-     * Deserializes a GerritTriggeredEvent from JSON.
-     *
-     * @param eventJson the JSON string
-     * @return deserialized event, or null if deserialization fails
-     */
-    private GerritTriggeredEvent deserializeEvent(String eventJson) {
-        try {
-            return GSON.fromJson(eventJson, GerritTriggeredEvent.class);
-        } catch (Exception e) {
-            logger.error("Failed to deserialize event from JSON", e);
-            return null;
-        }
-    }
-
-    /**
-     * Reconstructs a MemoryImprint from distributed data.
-     *
-     * @param event the event
-     * @param data the serialized data
-     * @return reconstructed MemoryImprint
-     */
-    private MemoryImprint reconstructMemoryImprint(GerritTriggeredEvent event, MemoryImprintData data) {
-        MemoryImprint imprint = new MemoryImprint(event);
-
-        if (data.getEntries() != null) {
-            Jenkins jenkins = Jenkins.getInstanceOrNull();
-            if (jenkins == null) {
-                logger.warn("Jenkins instance not available, cannot reconstruct MemoryImprint");
-                return imprint;
-            }
-
-            for (EntryData entryData : data.getEntries()) {
-                String projectFullName = entryData.getProjectFullName();
-                Job project = jenkins.getItemByFullName(projectFullName, Job.class);
-
-                if (project != null) {
-                    if (entryData.getBuildId() != null) {
-                        Run build = project.getBuild(entryData.getBuildId());
-                        if (build != null) {
-                            imprint.set(project, build, entryData.isBuildCompleted());
-                        } else {
-                            // Build not found, but project exists - add entry without build
-                            imprint.set(project);
-                        }
-                    } else {
-                        // No build ID - project triggered but not started
-                        imprint.set(project);
-                    }
-
-                    // Restore additional entry data
-                    Entry entry = imprint.getEntry(project);
-                    if (entry != null) {
-                        entry.setCancelled(entryData.isCancelled());
-                        entry.setCustomUrl(entryData.getCustomUrl());
-                        entry.setUnsuccessfulMessage(entryData.getUnsuccessfulMessage());
-                    }
-                }
-            }
-        }
-
-        return imprint;
-    }
-
-    /**
-     * Converts a MemoryImprint to serializable data.
-     *
-     * @param event the event
-     * @param imprint the memory imprint
-     * @return serializable data
-     */
-    private MemoryImprintData toMemoryImprintData(GerritTriggeredEvent event, MemoryImprint imprint) {
-        MemoryImprintData data = new MemoryImprintData();
-        data.setEventJson(serializeEvent(event));
-
-        for (Entry entry : imprint.getEntries()) {
-            Job project = entry.getProject();
-            if (project != null) {
-                EntryData entryData = new EntryData();
-                entryData.setProjectFullName(project.getFullName());
-
-                Run build = entry.getBuild();
-                if (build != null) {
-                    entryData.setBuildId(build.getId());
-                }
-
-                entryData.setBuildCompleted(entry.isBuildCompleted());
-                entryData.setCancelled(entry.isCancelled());
-                entryData.setCustomUrl(entry.getCustomUrl());
-                entryData.setUnsuccessfulMessage(entry.getUnsuccessfulMessage());
-                entryData.setTriggeredTimestamp(entry.getTriggeredTimestamp());
-                entryData.setCompletedTimestamp(entry.getCompletedTimestamp());
-                entryData.setStartedTimestamp(entry.getStartedTimestamp());
-
-                data.addEntry(entryData);
-            }
-        }
-
-        return data;
-    }
-
-    /**
-     * Writes a MemoryImprint back to distributed storage.
-     * Call this after modifying an imprint in cluster mode.
-     *
-     * @param event the event
-     * @param imprint the modified memory imprint
-     */
-    private void writeBackToDistributed(GerritTriggeredEvent event, MemoryImprint imprint) {
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = toMemoryImprintData(event, imprint);
-                map.put(key, data);
-                logger.trace("Wrote modified imprint back to distributed memory: {}", key);
-            }
-        }
-    }
 
     /**
      * Gets the memory of a specific event.
@@ -290,20 +112,6 @@ public class BuildMemory {
      * @return the memory.
      */
     public synchronized MemoryImprint getMemoryImprint(GerritTriggeredEvent event) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-                if (data != null) {
-                    return reconstructMemoryImprint(event, data);
-                }
-                return null;
-            }
-        }
-
-        // Fallback to local memory
         return memory.get(event);
     }
 
@@ -376,56 +184,6 @@ public class BuildMemory {
      * @param build the build.
      */
     public synchronized void completed(GerritTriggeredEvent event, Run build) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-
-                if (data == null) {
-                    // Shouldn't happen but just in case, keep the memory
-                    logger.debug("Build completed without being registered first (distributed mode).");
-                    data = new MemoryImprintData();
-                    data.setEventJson(serializeEvent(event));
-                }
-
-                // Update entry for this project as completed
-                String projectFullName = build.getParent().getFullName();
-                boolean found = false;
-
-                if (data.getEntries() != null) {
-                    for (EntryData entry : data.getEntries()) {
-                        if (projectFullName.equals(entry.getProjectFullName())) {
-                            if (entry.getBuildId() == null) {
-                                entry.setBuildId(build.getId());
-                            }
-                            entry.setBuildCompleted(true);
-                            entry.setCompletedTimestamp(System.currentTimeMillis());
-                            found = true;
-                            break;
-                        }
-                    }
-                }
-
-                if (!found) {
-                    // Entry not found, create new one as completed
-                    EntryData entryData = new EntryData();
-                    entryData.setProjectFullName(projectFullName);
-                    entryData.setBuildId(build.getId());
-                    entryData.setBuildCompleted(true);
-                    entryData.setCompletedTimestamp(System.currentTimeMillis());
-                    data.addEntry(entryData);
-                }
-
-                // Store back to Hazelcast
-                map.put(key, data);
-                logger.trace("Build completed event stored in distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         MemoryImprint pb = memory.get(event);
         if (pb == null) {
             //Shoudn't happen but just in case, keep the memory.
@@ -442,51 +200,6 @@ public class BuildMemory {
      * @param build the build.
      */
     public synchronized void started(GerritTriggeredEvent event, Run build) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-
-                if (data == null) {
-                    logger.warn("Build started without being registered first (distributed mode).");
-                    data = new MemoryImprintData();
-                    data.setEventJson(serializeEvent(event));
-                }
-
-                // Update entry for this project with build info
-                String projectFullName = build.getParent().getFullName();
-                boolean found = false;
-
-                if (data.getEntries() != null) {
-                    for (EntryData entry : data.getEntries()) {
-                        if (projectFullName.equals(entry.getProjectFullName())) {
-                            entry.setBuildId(build.getId());
-                            entry.setStartedTimestamp(System.currentTimeMillis());
-                            found = true;
-                            break;
-                        }
-                    }
-                }
-
-                if (!found) {
-                    // Entry not found, create new one
-                    EntryData entryData = new EntryData();
-                    entryData.setProjectFullName(projectFullName);
-                    entryData.setBuildId(build.getId());
-                    entryData.setStartedTimestamp(System.currentTimeMillis());
-                    data.addEntry(entryData);
-                }
-
-                // Store back to Hazelcast
-                map.put(key, data);
-                logger.trace("Build started event stored in distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         MemoryImprint pb = memory.get(event);
         if (pb == null) {
             //A build should not start for a job that hasn't been registered. Keep the memory anyway.
@@ -504,32 +217,6 @@ public class BuildMemory {
      * @param project the project that was triggered.
      */
     public synchronized void triggered(GerritTriggeredEvent event, Job project) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-
-                if (data == null) {
-                    // Create new memory imprint data
-                    data = new MemoryImprintData();
-                    data.setEventJson(serializeEvent(event));
-                }
-
-                // Add entry for triggered project
-                EntryData entryData = new EntryData();
-                entryData.setProjectFullName(project.getFullName());
-                data.addEntry(entryData);
-
-                // Store back to Hazelcast
-                map.put(key, data);
-                logger.trace("Triggered event stored in distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         MemoryImprint pb = memory.get(event);
         if (pb == null) {
             pb = new MemoryImprint(event);
@@ -551,63 +238,6 @@ public class BuildMemory {
             GerritTriggeredEvent event,
             Job project,
             List<Run> otherBuilds) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-
-                if (data == null) {
-                    // Create new memory imprint data
-                    data = new MemoryImprintData();
-                    data.setEventJson(serializeEvent(event));
-
-                    if (otherBuilds != null) {
-                        // Populate with old build info
-                        for (Run build : otherBuilds) {
-                            EntryData entryData = new EntryData();
-                            entryData.setProjectFullName(build.getParent().getFullName());
-                            entryData.setBuildId(build.getId());
-                            entryData.setBuildCompleted(!build.isBuilding());
-                            data.addEntry(entryData);
-                        }
-                    }
-                }
-
-                // Reset the retriggered project (clear build info)
-                String projectFullName = project.getFullName();
-                boolean found = false;
-
-                if (data.getEntries() != null) {
-                    for (EntryData entry : data.getEntries()) {
-                        if (projectFullName.equals(entry.getProjectFullName())) {
-                            // Reset this entry
-                            entry.setBuildId(null);
-                            entry.setBuildCompleted(false);
-                            entry.setStartedTimestamp(null);
-                            entry.setCompletedTimestamp(null);
-                            found = true;
-                            break;
-                        }
-                    }
-                }
-
-                if (!found) {
-                    // Add new entry for retriggered project
-                    EntryData entryData = new EntryData();
-                    entryData.setProjectFullName(projectFullName);
-                    data.addEntry(entryData);
-                }
-
-                // Store back to Hazelcast
-                map.put(key, data);
-                logger.trace("Retriggered event stored in distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         MemoryImprint pb = memory.get(event);
         if (pb == null) {
             pb = new MemoryImprint(event);
@@ -629,51 +259,6 @@ public class BuildMemory {
      * @param project     the project that has been retriggered.
      */
     public synchronized void cancelled(GerritTriggeredEvent event, Job project) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = map.get(key);
-
-                if (data == null) {
-                    // Shouldn't happen but just in case, keep the memory
-                    data = new MemoryImprintData();
-                    data.setEventJson(serializeEvent(event));
-                }
-
-                // Find or create entry for this project
-                String projectFullName = project.getFullName();
-                boolean found = false;
-
-                if (data.getEntries() != null) {
-                    for (EntryData entry : data.getEntries()) {
-                        if (projectFullName.equals(entry.getProjectFullName())) {
-                            entry.setCancelled(true);
-                            entry.setBuildCompleted(true);
-                            found = true;
-                            break;
-                        }
-                    }
-                }
-
-                if (!found) {
-                    // Create new entry as cancelled
-                    EntryData entryData = new EntryData();
-                    entryData.setProjectFullName(projectFullName);
-                    entryData.setCancelled(true);
-                    entryData.setBuildCompleted(true);
-                    data.addEntry(entryData);
-                }
-
-                // Store back to Hazelcast
-                map.put(key, data);
-                logger.trace("Cancelled event stored in distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         MemoryImprint pb = memory.get(event);
         if (pb == null) {
             //Shoudn't happen but just in case, keep the memory.
@@ -693,18 +278,6 @@ public class BuildMemory {
      * @param event the event.
      */
     public synchronized void forget(GerritTriggeredEvent event) {
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                map.remove(key);
-                logger.trace("Forgot event from distributed memory: {}", key);
-                return;
-            }
-        }
-
-        // Fallback to local memory
         memory.remove(event);
     }
 
@@ -871,9 +444,6 @@ public class BuildMemory {
             if (entry != null) {
                 logger.trace("Recording custom URL for {}: {}", event, customUrl);
                 entry.setCustomUrl(customUrl);
-
-                // Write back to distributed storage if in cluster mode
-                writeBackToDistributed(event, pb);
             }
         }
     }
@@ -894,9 +464,6 @@ public class BuildMemory {
             if (entry != null) {
                 logger.trace("Recording unsuccessful message for {}: {}", event, unsuccessfulMessage);
                 entry.setUnsuccessfulMessage(unsuccessfulMessage);
-
-                // Write back to distributed storage if in cluster mode
-                writeBackToDistributed(event, pb);
             }
         }
     }
@@ -921,33 +488,6 @@ public class BuildMemory {
      */
     public synchronized void removeProject(Job project) {
         String projectFullName = project.getFullName();
-
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                // Iterate over all entries in distributed memory
-                for (Map.Entry<BuildMemoryKey, MemoryImprintData> mapEntry : map.entrySet()) {
-                    MemoryImprintData data = mapEntry.getValue();
-                    if (data.getEntries() != null) {
-                        // Remove entries matching this project
-                        boolean removed = data.getEntries().removeIf(
-                            entry -> projectFullName.equals(entry.getProjectFullName())
-                        );
-
-                        // If we removed anything, update the map
-                        if (removed) {
-                            map.put(mapEntry.getKey(), data);
-                            logger.trace("Removed project {} from distributed memory entry: {}",
-                                projectFullName, mapEntry.getKey());
-                        }
-                    }
-                }
-                return;
-            }
-        }
-
-        // Fallback to local memory
         for (MemoryImprint memoryImprint : memory.values()) {
             memoryImprint.removeProject(projectFullName);
         }
@@ -961,30 +501,6 @@ public class BuildMemory {
     @NonNull
     public synchronized BuildMemoryReport report() {
         BuildMemoryReport report = new BuildMemoryReport();
-
-        // Try distributed memory first if cluster mode enabled
-        if (isClusterModeEnabled()) {
-            IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-            if (map != null) {
-                // Read all entries from distributed memory
-                for (Map.Entry<BuildMemoryKey, MemoryImprintData> mapEntry : map.entrySet()) {
-                    MemoryImprintData data = mapEntry.getValue();
-                    GerritTriggeredEvent event = deserializeEvent(data.getEventJson());
-
-                    if (event != null) {
-                        MemoryImprint imprint = reconstructMemoryImprint(event, data);
-                        List<Entry> triggered = new LinkedList<Entry>();
-                        for (Entry tr : imprint.list) {
-                            triggered.add(tr.clone());
-                        }
-                        report.put(event, triggered);
-                    }
-                }
-                return report;
-            }
-        }
-
-        // Fallback to local memory
         for (Map.Entry<GerritTriggeredEvent, MemoryImprint> entry : memory.entrySet()) {
             List<Entry> triggered = new LinkedList<Entry>();
             for (Entry tr : entry.getValue().list) {
@@ -995,83 +511,6 @@ public class BuildMemory {
         return report;
     }
 
-    /**
-     * Migrates data from local memory to distributed Hazelcast storage.
-     * Called when cluster mode is enabled.
-     * <p>
-     * This copies all entries from the local TreeMap to the Hazelcast IMap.
-     */
-    public synchronized void migrateToDistributed() {
-        if (!isClusterModeEnabled()) {
-            logger.warn("Cannot migrate to distributed: cluster mode is not enabled");
-            return;
-        }
-
-        IMap<BuildMemoryKey, MemoryImprintData> map = getDistributedMemory();
-        if (map == null) {
-            logger.error("Cannot migrate to distributed: Hazelcast map is not available");
-            return;
-        }
-
-        logger.info("Starting migration from local to distributed BuildMemory...");
-        int migratedCount = 0;
-
-        try {
-            for (Map.Entry<GerritTriggeredEvent, MemoryImprint> entry : memory.entrySet()) {
-                GerritTriggeredEvent event = entry.getKey();
-                MemoryImprint imprint = entry.getValue();
-
-                BuildMemoryKey key = new BuildMemoryKey(event);
-                MemoryImprintData data = toMemoryImprintData(event, imprint);
-
-                map.put(key, data);
-                migratedCount++;
-            }
-
-            logger.info("Successfully migrated {} BuildMemory entries to distributed storage", migratedCount);
-        } catch (Exception e) {
-            logger.error("Error during BuildMemory migration to distributed storage", e);
-        }
-    }
-
-    /**
-     * Migrates data from distributed Hazelcast storage to local memory.
-     * Called when cluster mode is disabled.
-     * <p>
-     * This copies all entries from the Hazelcast IMap to the local TreeMap.
-     */
-    public synchronized void migrateToLocal() {
-        IMap<BuildMemoryKey, MemoryImprintData> map = distributedMemory;
-        if (map == null) {
-            logger.debug("No distributed memory to migrate from");
-            return;
-        }
-
-        logger.info("Starting migration from distributed to local BuildMemory...");
-        int migratedCount = 0;
-
-        try {
-            for (Map.Entry<BuildMemoryKey, MemoryImprintData> mapEntry : map.entrySet()) {
-                MemoryImprintData data = mapEntry.getValue();
-                GerritTriggeredEvent event = deserializeEvent(data.getEventJson());
-
-                if (event != null) {
-                    MemoryImprint imprint = reconstructMemoryImprint(event, data);
-                    memory.put(event, imprint);
-                    migratedCount++;
-                } else {
-                    logger.warn("Failed to deserialize event during migration: {}", mapEntry.getKey());
-                }
-            }
-
-            logger.info("Successfully migrated {} BuildMemory entries to local storage", migratedCount);
-
-            // Clear distributed memory reference
-            distributedMemory = null;
-        } catch (Exception e) {
-            logger.error("Error during BuildMemory migration to local storage", e);
-        }
-    }
 
     /**
      * A holder for all builds triggered by one event.
@@ -1180,7 +619,7 @@ public class BuildMemory {
          * @param build          the build
          * @param buildCompleted if the build is finished.
          */
-        private synchronized void set(Job project, Run build, boolean buildCompleted) {
+        synchronized void set(Job project, Run build, boolean buildCompleted) {
             Entry entry = getEntry(project);
             if (entry == null) {
                 entry = new Entry(project, build);
@@ -1259,7 +698,7 @@ public class BuildMemory {
          * @param project the project.
          * @return the entry or null if nothing is found.
          */
-        private Entry getEntry(@NonNull Job project) {
+        Entry getEntry(@NonNull Job project) {
             for (Entry entry : list) {
                 if (entry != null && project.equals(entry.getProject())) {
                     return entry;
@@ -1557,7 +996,7 @@ public class BuildMemory {
              *
              * @param customUrl the URL.
              */
-            private void setCustomUrl(String customUrl) {
+            void setCustomUrl(String customUrl) {
                 this.customUrl = customUrl;
             }
 
@@ -1575,7 +1014,7 @@ public class BuildMemory {
              *
              * @param unsuccessfulMessage the message.
              */
-            private void setUnsuccessfulMessage(String unsuccessfulMessage) {
+            void setUnsuccessfulMessage(String unsuccessfulMessage) {
                 this.unsuccessfulMessage = unsuccessfulMessage;
             }
 
@@ -1602,7 +1041,7 @@ public class BuildMemory {
              *
              * @param buildCompleted true if the build is completed.
              */
-            private void setBuildCompleted(boolean buildCompleted) {
+            void setBuildCompleted(boolean buildCompleted) {
                 this.buildCompleted = buildCompleted;
                 if (buildCompleted) {
                     this.completedTimestamp = System.currentTimeMillis();
@@ -1622,7 +1061,7 @@ public class BuildMemory {
              *
              * @param cancelled true if the build was cancelled while in the queue.
              */
-            private void setCancelled(boolean cancelled) {
+            void setCancelled(boolean cancelled) {
                 this.cancelled = cancelled;
             }
